+++ NEWS.md
@@ -1,3 +1,9 @@
+# NPSutils 0.2.0.1
+  
+  * updated get_data_package to retrieve 1 or more files from a given reference, if for instance a data package has multiple files associated with it. get_data_package is file extension agnostic.
+  * get_data_package will check whether any files are .zip and attempt to extract them.
+
+
 # NPSutils 0.2.0.0
 
 Facelift to the entire package: